requirements.txt
fastapi==0.89.0
strawberry-graphql==0.151.2
uvicorn==0.20.0
pyyaml
